feat: PC (16:9) template bank + verify-after-fire wiring + Default-trait migrations
Consensus across 6 gates (reliability/performance/extensibility/governance/security/integration): 5 GO + 1 CONDITIONAL (governance), no critical or code-quality blockers.

Production code changes are limited to: (1) idiomatic refactors (let-chains, range.contains, .is_empty, redundant-cast removal, #[derive(Default)]), (2) Default-trait migration of RecoveryPolicy/StrategyRegistry/TemplateStore/InputMethod/EngineKind — source-compatible and each pinned by an equivalence regression test locking max_retries==5 and empty-registry invariants, (3) end-to-end wiring of the pre-existing verify-after-fire path (PipelineDriver::with_verify, default true) through the CLI Run command and 4 runner adapters.

The +592-line pipeline.rs / +73-line scale.rs additions are entirely inside #[cfg(test)] — PC-bank validators, namespace-collision assertions, and the 1258x708 ROI invariant. The 20:9/16:9 coexistence invariant is regression-tested.

Verification evidence:
- cargo check --all: clean (0.58s-1.53s, exit 0, zero warnings)
- cargo clippy --all-targets -- -D warnings: zero warnings
- cargo nextest run --workspace: 196 passed, 0 failed (20.285s)
- No new library unwrap/expect/panic; all new panic paths are #[cfg(test)] with the 3-point #[allow] trio
- No Cargo.toml changes; supply-chain surface unchanged
- No new unsafe; Win32 FFI blocks pre-existing and only cast-cleaned
- Recovery state machine non-recursive, GiveUp boundary preserved

Deferred to Release phase (non-blocking):
- Split single working tree into logical commits per 1-commit=1-change rule: (a) Default-trait migrations + tests, (b) PC scene/pipeline templates + cfg(test) validators + docs/pc-capture-dimensions.md, (c) verify_after_fire flag + CLI wiring, (d) fmt/let-chains mechanical refactors
- Add changelog note that Run now verifies by default (--verify-after-fire=false to restore old behavior)
- Confirm untracked artifacts (analyze_menu_bar.rs, tests/, menu_pc/) are intentionally staged or excluded
- Pre-existing template_store.rs unwrap/panic surface remains tracked as task #34 (not this commit)

+++ crates/anaden-device/src/win32_capture.rs
@@ -3,7 +3,7 @@
 //!
 //! `probe_windows_capture.rs` で動作検証済みの PrintWindow(PW_RENDERFULLCONTENT=0x2)
 //! + GDI 連鎖(GetDIBits/BGRA→RGBA)をそのまま活かす。BitBlt 単体は cocos2d-x OpenGL
-//! 描画を拾えず黒画像になるため使用しない(プローブ実証済み)。
+//!   描画を拾えず黒画像になるため使用しない(プローブ実証済み)。
 //!
 //! 概要:
 //! - `Win32Capture::new("AnotherEden.exe")` で生成(DPI アウェア化を冪等に実施)。
@@ -400,7 +400,7 @@ pub(crate) fn capture_via_printwindow(hwnd: HWND, w: u32, h: u32) -> Result<Vec<
             mem_dc,
             bmp,
             0,
-            h as u32,
+            h,
             Some(pixels.as_mut_ptr() as *mut _),
             &mut bi,
             DIB_RGB_COLORS,
